==> notebooks/05_SVM.ipynb <==
import pandas as pd
import pickle
from sklearn.metrics import confusion_matrix, roc_curve, roc_auc_score, accuracy_score, classification_report
from sklearn.feature_selection import SelectFromModel, RFE
from sklearn import linear_model
from imblearn.pipeline import Pipeline
from imblearn.over_sampling import SMOTE 
from sklearn.svm import SVC

def evaluate_model(ytest, ypred, ypred_proba = None):
    if ypred_proba is not None:
        print('ROC-AUC score of the model: {}'.format(roc_auc_score(ytest, ypred_proba[:, 1])))
    print('Accuracy of the model: {}\n'.format(accuracy_score(ytest, ypred)))
    print('Classification report: \n{}\n'.format(classification_report(ytest, ypred)))
    print('Confusion matrix: \n{}\n'.format(confusion_matrix(ytest, ypred)))

X_train = pd.read_parquet("../data/processed/X_train.parquet")
y_train = pd.read_parquet("../data/processed/y_train.parquet")['isfraud']
X_test = pd.read_parquet("../data/processed/X_test.parquet")
y_test = pd.read_parquet("../data/processed/y_test.parquet")['isfraud']
X_train_scaled = pd.read_parquet("../data/processed/X_train_scaled.parquet")
X_test_scaled = pd.read_parquet("../data/processed/X_test_scaled.parquet")

classifier = SVC(kernel='linear', probability=True)
smt = SMOTE(sampling_strategy=0.5, n_jobs=-1)

clf_svm = Pipeline(steps=[
    ('smote', smt),
    ('clasificador', classifier)])

clf_svm.fit(X_train_scaled, y_train)

y_pred_svm = clf_svm.predict(X_test_scaled)
ypred_proba_svm = clf_svm.predict_proba(X_test_scaled)

evaluate_model(y_test, y_pred_svm, ypred_proba_svm)

Unfortunately, we have not been able to run the model. The fit part has never loaded. We have tried it for several days and with different programs but we have not been able to obtain a prediction. It's a more complex model and might have been good for our problem.
